webapp/backend/requirements.txt
beautifulsoup4
django-apscheduler
django
djangorestframework
lxml
requests
django-cors-headers